Move some bits to internal/
Move some of the OS abstractions, parsing, and other bits to internal/
package.

Also split out conn.go to rows.go and stmt.go.
