+++ connector_test.go
@@ -9,6 +9,8 @@ import (
 	"os"
 	"reflect"
 	"testing"
+
+	"github.com/lib/pq/internal/pqtest"
 )
 
 func TestNewConnector_WorksWithOpenDB(t *testing.T) {
@@ -128,5 +130,37 @@ func TestIsUTF8(t *testing.T) {
 		if g := isUTF8(test.name); g != test.want {
 			t.Errorf("isUTF8(%q) = %v want %v", test.name, g, test.want)
 		}
+	}
+}
+
+func TestParseURL(t *testing.T) {
+	tests := []struct {
+		in, want, wantErr string
+	}{
+		{"postgres://", "", ""},
+		{"postgres://hostname.remote", "host='hostname.remote'", ""},
+		{"postgres://[::1]:1234", "host='::1' port='1234'", ""},
+		{"postgres://username:top%20secret@hostname.remote:1234/database",
+			`dbname='database' host='hostname.remote' password='top secret' port='1234' user='username'`, ""},
+		{"postgres://localhost/a%2Fb", "dbname='a/b' host='localhost'", ""},
+
+		{"", "", "invalid connection protocol:"},
+		{"http://hostname.remote", "", "invalid connection protocol: http"},
+
+		//{"postgresql://%2Fvar%2Flib%2Fpostgresql/dbname", "", ``},
+		//{"postgres:// host/db", "dbname='db' host='host'", ""},
+		//{"postgres://host/db ", "dbname='db' host='host'", ""},
+	}
+
+	for _, tt := range tests {
+		t.Run("", func(t *testing.T) {
+			have, err := ParseURL(tt.in)
+			if !pqtest.ErrorContains(err, tt.wantErr) {
+				t.Fatal(err)
+			}
+			if have != tt.want {
+				t.Errorf("\nhave: %q\nwant: %q", have, tt.want)
+			}
+		})
 	}
 }
