+++ internal/pqutil/perm_unsupported.go
@@ -1,0 +1,5 @@
+//go:build windows || plan9
+
+package pqutil
+
+func sslKeyPermissions(string) error { return nil }
