+++ oid/doc.go
@@ -1,5 +1,6 @@
-// Package oid contains OID constants
-// as defined by the Postgres server.
+//go:generate go run ./gen.go
+
+// Package oid contains OID constants as defined by the Postgres server.
 package oid
 
 // Oid is a Postgres Object ID.
