--- user_posix.go
@@ -1,22 +1,0 @@
-//go:build aix || darwin || dragonfly || freebsd || (linux && !android && !appengine) || nacl || netbsd || openbsd || plan9 || solaris || rumprun || illumos
-
-package pq
-
-import (
-	"os"
-	"os/user"
-)
-
-func userCurrent() (string, error) {
-	u, err := user.Current()
-	if err == nil {
-		return u.Username, nil
-	}
-
-	name := os.Getenv("USER")
-	if name != "" {
-		return name, nil
-	}
-
-	return "", ErrCouldNotDetectUsername
-}
